move to geometry and add docs
