+++ kornia/geometry/camera/__init__.py
@@ -1,5 +1,28 @@
+from .distortion_affine import distort_points_affine, dx_distort_points_affine, undistort_points_affine
 from .perspective import project_points, unproject_points
 from .pinhole import PinholeCamera, cam2pixel, pixel2cam
+from .projection_orthographic import (
+    dx_project_points_orthographic,
+    project_points_orthographic,
+    unproject_points_orthographic,
+)
+from .projection_z1 import dx_project_points_z1, project_points_z1, unproject_points_z1
 from .stereo import StereoCamera
 
-__all__ = ["PinholeCamera", "StereoCamera", "pixel2cam", "cam2pixel", "unproject_points", "project_points"]
+__all__ = [
+    "PinholeCamera",
+    "StereoCamera",
+    "cam2pixel",
+    "distort_points_affine",
+    "dx_distort_points_affine",
+    "dx_project_points_orthographic",
+    "dx_project_points_z1",
+    "pixel2cam",
+    "project_points",
+    "project_points_orthographic",
+    "project_points_z1",
+    "unproject_points",
+    "unproject_points_orthographic",
+    "unproject_points_z1",
+    "undistort_points_affine",
+]
