+++ test/geometry/camera/test_distortion.py
@@ -1,0 +1,111 @@
+import pytest
+import torch
+
+from kornia.geometry.camera.distortion_affine import (
+    distort_points_affine,
+    dx_distort_points_affine,
+    undistort_points_affine,
+)
+from kornia.testing import BaseTester, tensor_to_gradcheck_var
+
+
+class TestDistortionAffine(BaseTester):
+    def test_smoke(self, device, dtype):
+        points = torch.tensor([1.0, 2.0], device=device, dtype=dtype)
+        params = torch.tensor([600.0, 600.0, 319.5, 239.5], device=device, dtype=dtype)
+        assert distort_points_affine(points, params) is not None
+
+    def _test_cardinality_distort_batch(self, device, dtype, batch_size):
+        batch_tuple = (batch_size,) if batch_size is not None else ()
+        points = torch.rand(batch_tuple + (2,), device=device, dtype=dtype)
+        params = torch.rand(batch_tuple + (4,), device=device, dtype=dtype)
+        assert distort_points_affine(points, params).shape == batch_tuple + (2,)
+
+    def _test_cardinality_undistort_batch(self, device, dtype, batch_size):
+        batch_tuple = (batch_size,) if batch_size is not None else ()
+        points = torch.rand(batch_tuple + (2,), device=device, dtype=dtype)
+        params = torch.rand(batch_tuple + (4,), device=device, dtype=dtype)
+        assert undistort_points_affine(points, params).shape == batch_tuple + (2,)
+
+    @pytest.mark.parametrize("batch_size", [None, 1, 2, 3])
+    def test_cardinality(self, device, dtype, batch_size):
+        self._test_cardinality_distort_batch(device, dtype, batch_size)
+        self._test_cardinality_undistort_batch(device, dtype, batch_size)
+
+    def test_distort_points_affine(self, device, dtype):
+        points = torch.tensor([1.0, 2.0], device=device, dtype=dtype)
+        params = torch.tensor([600.0, 600.0, 319.5, 239.5], device=device, dtype=dtype)
+        expected = torch.tensor([919.5000, 1439.5000], device=device, dtype=dtype)
+        self.assert_close(distort_points_affine(points, params), expected)
+
+    def test_distort_points_affine_batch(self, device, dtype):
+        points = torch.tensor([[1.0, 2.0], [3.0, 4.0]], device=device, dtype=dtype)
+        params = torch.tensor([[600.0, 600.0, 319.5, 239.5], [600.0, 600.0, 319.5, 239.5]], device=device, dtype=dtype)
+        expected = torch.tensor([[919.5000, 1439.5000], [2119.5000, 2639.5000]], device=device, dtype=dtype)
+        self.assert_close(distort_points_affine(points, params), expected)
+
+    def test_undistort_points_affine(self, device, dtype):
+        points = torch.tensor([601.0, 602.0], device=device, dtype=dtype)
+        params = torch.tensor([600.0, 600.0, 319.5, 239.5], device=device, dtype=dtype)
+        expected = torch.tensor([0.4692, 0.6042], device=device, dtype=dtype)
+        self.assert_close(undistort_points_affine(points, params), expected)
+
+    def test_undistort_points_affine_batch(self, device, dtype):
+        points = torch.tensor([[601.0, 602.0], [1203.0, 1204.0]], device=device, dtype=dtype)
+        params = torch.tensor([[600.0, 600.0, 319.5, 239.5], [600.0, 600.0, 319.5, 239.5]], device=device, dtype=dtype)
+        expected = torch.tensor([[0.4692, 0.6042], [1.4725, 1.6075]], device=device, dtype=dtype)
+        self.assert_close(undistort_points_affine(points, params), expected)
+
+    def test_dx_distort_points_affine(self, device, dtype):
+        points = torch.tensor([1.0, 2.0], device=device, dtype=dtype)
+        params = torch.tensor([600.0, 600.0, 319.5, 239.5], device=device, dtype=dtype)
+        expected = torch.tensor([[600.0, 0.0], [0.0, 600.0]], device=device, dtype=dtype)
+        self.assert_close(dx_distort_points_affine(points, params), expected)
+
+    def test_exception(self, device, dtype) -> None:
+        points = torch.tensor([1.0, 2.0], device=device, dtype=dtype)
+        params = torch.tensor([600.0, 600.0, 319.5], device=device, dtype=dtype)
+        with pytest.raises(TypeError):
+            distort_points_affine(points, params)
+
+    def _test_gradcheck_distort(self, device):
+        points = torch.tensor([1.0, 2.0], device=device, dtype=torch.float64)
+        points = tensor_to_gradcheck_var(points)
+        params = torch.tensor([600.0, 600.0, 319.5, 239.5], device=device, dtype=torch.float64)
+        params = tensor_to_gradcheck_var(params)
+        self.gradcheck(distort_points_affine, (points, params))
+
+    def _test_gradcheck_undistort(self, device):
+        points = torch.tensor([601.0, 602.0], device=device, dtype=torch.float64)
+        points = tensor_to_gradcheck_var(points)
+        params = torch.tensor([600.0, 600.0, 319.5, 239.5], device=device, dtype=torch.float64)
+        params = tensor_to_gradcheck_var(params)
+        self.gradcheck(undistort_points_affine, (points, params))
+
+    def test_gradcheck(self, device) -> None:
+        self._test_gradcheck_distort(device)
+        self._test_gradcheck_undistort(device)
+
+    def _test_jit_distort(self, device, dtype) -> None:
+        points = torch.tensor([1.0, 2.0], device=device, dtype=dtype)
+        params = torch.tensor([600.0, 600.0, 319.5, 239.5], device=device, dtype=dtype)
+        op_script = torch.jit.script(distort_points_affine)
+        actual = op_script(points, params)
+        expected = distort_points_affine(points, params)
+        self.assert_close(actual, expected)
+
+    def _test_jit_undistort(self, device, dtype) -> None:
+        points = torch.tensor([601.0, 602.0], device=device, dtype=dtype)
+        params = torch.tensor([600.0, 600.0, 319.5, 239.5], device=device, dtype=dtype)
+        op_script = torch.jit.script(undistort_points_affine)
+        actual = op_script(points, params)
+        expected = undistort_points_affine(points, params)
+        self.assert_close(actual, expected)
+
+    def test_jit(self, device, dtype) -> None:
+        self._test_jit_distort(device, dtype)
+        self._test_jit_undistort(device, dtype)
+
+    @pytest.mark.skip(reason="Unnecessary test")
+    def test_module(self, device, dtype) -> None:
+        pass
